requirements.txt
requests
pyyaml
pytest
pytest-cov
pytest-flakes
pytest-pep8
mock
